Move methods from AuthController.java to LoginApiController and RegisterApiController Delete AuthController delete one RegisterReq delete one LoginReq
