+++ src/main/java/ru/skypro/homework/controller/LoginApiController.java
@@ -3,6 +3,8 @@ package ru.skypro.homework.controller;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.enums.ParameterIn;
 import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import ru.skypro.homework.controller.api.LoginApi;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -13,13 +15,14 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.RequestBody;
 import ru.skypro.homework.dto.LoginReq;
+import ru.skypro.homework.service.AuthService;
 
 import javax.validation.Valid;
 import javax.servlet.http.HttpServletRequest;
 import java.io.IOException;
 
 @javax.annotation.Generated(value = "ru.skypro.homeworkcodegen.v3.generators.java.SpringCodegen", date = "2023-02-06T18:24:36.081075022Z[GMT]")
-@RestController
+@RestController("login")
 @CrossOrigin(value = "http://localhost:3000")
 public class LoginApiController implements LoginApi {
 
@@ -29,20 +32,21 @@ public class LoginApiController implements LoginApi {
 
     private final HttpServletRequest request;
 
-    @org.springframework.beans.factory.annotation.Autowired
-    public LoginApiController(ObjectMapper objectMapper, HttpServletRequest request) {
+    private final AuthService authService;
+
+    public LoginApiController(ObjectMapper objectMapper, HttpServletRequest request, AuthService authService) {
         this.objectMapper = objectMapper;
         this.request = request;
+        this.authService = authService;
     }
 
     public ResponseEntity<Object> login(@Parameter(in = ParameterIn.DEFAULT, description = "", required=true, schema=@Schema()) @Valid @RequestBody LoginReq body) {
         String accept = request.getHeader("Accept");
         if (accept != null && accept.contains("application/json")) {
-            try {
-                return new ResponseEntity<Object>(objectMapper.readValue("{ }", Object.class), HttpStatus.NOT_IMPLEMENTED);
-            } catch (IOException e) {
-                log.error("Couldn't serialize response for content type application/json", e);
-                return new ResponseEntity<Object>(HttpStatus.INTERNAL_SERVER_ERROR);
+            if (authService.login(body.getUsername(), body.getPassword())) {
+                return ResponseEntity.ok().build();
+            } else {
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
             }
         }
