+++ src/main/java/ru/skypro/homework/controller/RegisterApiController.java
@@ -1,5 +1,7 @@
 package ru.skypro.homework.controller;
 
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import ru.skypro.homework.controller.api.RegisterApi;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -13,12 +15,16 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.RequestBody;
 import ru.skypro.homework.dto.RegisterReq;
+import ru.skypro.homework.dto.Role;
+import ru.skypro.homework.service.AuthService;
 
 import javax.validation.Valid;
 import javax.servlet.http.HttpServletRequest;
 
+import static ru.skypro.homework.dto.Role.USER;
+
 @javax.annotation.Generated(value = "ru.skypro.homeworkcodegen.v3.generators.java.SpringCodegen", date = "2023-02-06T18:24:36.081075022Z[GMT]")
-@RestController
+@RestController("register")
 @CrossOrigin(value = "http://localhost:3000")
 public class RegisterApiController implements RegisterApi {
 
@@ -27,16 +33,22 @@ public class RegisterApiController implements RegisterApi {
     private final ObjectMapper objectMapper;
 
     private final HttpServletRequest request;
+    private final AuthService authService;
 
-    @org.springframework.beans.factory.annotation.Autowired
-    public RegisterApiController(ObjectMapper objectMapper, HttpServletRequest request) {
+    public RegisterApiController(ObjectMapper objectMapper, HttpServletRequest request, AuthService authService) {
         this.objectMapper = objectMapper;
         this.request = request;
+        this.authService = authService;
     }
 
     public ResponseEntity<Void> register(@Parameter(in = ParameterIn.DEFAULT, description = "", required=true, schema=@Schema()) @Valid @RequestBody RegisterReq body) {
         String accept = request.getHeader("Accept");
-        return new ResponseEntity<Void>(HttpStatus.NOT_IMPLEMENTED);
+        Role role = body.getRole() == null ? USER : body.getRole();
+        if (authService.register(body, role)) {
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }
     }
 
 }
