+++ src/main/java/ru/skypro/homework/dto/RegisterReq.java
@@ -1,13 +1,184 @@
 package ru.skypro.homework.dto;
 
-import lombok.Data;
+import java.util.Objects;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import io.swagger.v3.oas.annotations.media.Schema;
+import org.springframework.validation.annotation.Validated;
 
-@Data
-public class RegisterReq {
-    private String username;
-    private String password;
-    private String firstName;
-    private String lastName;
-    private String phone;
-    private Role role;
+/**
+ * RegisterReq
+ */
+@Validated
+@javax.annotation.Generated(value = "ru.skypro.homeworkcodegen.v3.generators.java.SpringCodegen", date = "2023-02-06T18:24:36.081075022Z[GMT]")
+
+
+public class RegisterReq   {
+  @JsonProperty("username")
+  private String username = null;
+
+  @JsonProperty("password")
+  private String password = null;
+
+  @JsonProperty("firstName")
+  private String firstName = null;
+
+  @JsonProperty("lastName")
+  private String lastName = null;
+
+  @JsonProperty("phone")
+  private String phone = null;
+
+  private Role role;
+
+
+  public RegisterReq username(String username) {
+    this.username = username;
+    return this;
+  }
+
+  public void setRole(Role role) {
+    this.role = role;
+  }
+
+  /**
+   * Get username
+   * @return username
+   **/
+  @Schema(description = "")
+  
+    public String getUsername() {
+    return username;
+  }
+
+  public void setUsername(String username) {
+    this.username = username;
+  }
+
+  public RegisterReq password(String password) {
+    this.password = password;
+    return this;
+  }
+
+  /**
+   * Get password
+   * @return password
+   **/
+  @Schema(description = "")
+  
+    public String getPassword() {
+    return password;
+  }
+
+  public void setPassword(String password) {
+    this.password = password;
+  }
+
+  public RegisterReq firstName(String firstName) {
+    this.firstName = firstName;
+    return this;
+  }
+
+  /**
+   * Get firstName
+   * @return firstName
+   **/
+  @Schema(description = "")
+  
+    public String getFirstName() {
+    return firstName;
+  }
+
+  public void setFirstName(String firstName) {
+    this.firstName = firstName;
+  }
+
+  public RegisterReq lastName(String lastName) {
+    this.lastName = lastName;
+    return this;
+  }
+
+  /**
+   * Get lastName
+   * @return lastName
+   **/
+  @Schema(description = "")
+  
+    public String getLastName() {
+    return lastName;
+  }
+
+  public void setLastName(String lastName) {
+    this.lastName = lastName;
+  }
+
+  public RegisterReq phone(String phone) {
+    this.phone = phone;
+    return this;
+  }
+
+  /**
+   * Get phone
+   * @return phone
+   **/
+  @Schema(description = "")
+  
+    public String getPhone() {
+    return phone;
+  }
+
+  public void setPhone(String phone) {
+    this.phone = phone;
+  }
+
+  public Role getRole() {
+    return role;
+  }
+
+  @Override
+  public boolean equals(java.lang.Object o) {
+    if (this == o) {
+      return true;
+    }
+    if (o == null || getClass() != o.getClass()) {
+      return false;
+    }
+    RegisterReq registerReq = (RegisterReq) o;
+    return Objects.equals(this.username, registerReq.username) &&
+        Objects.equals(this.password, registerReq.password) &&
+        Objects.equals(this.firstName, registerReq.firstName) &&
+        Objects.equals(this.lastName, registerReq.lastName) &&
+        Objects.equals(this.phone, registerReq.phone) &&
+        Objects.equals(this.role, registerReq.role);
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(username, password, firstName, lastName, phone, role);
+  }
+
+  @Override
+  public String toString() {
+    StringBuilder sb = new StringBuilder();
+    sb.append("class RegisterReq {\n");
+    
+    sb.append("    username: ").append(toIndentedString(username)).append("\n");
+    sb.append("    password: ").append(toIndentedString(password)).append("\n");
+    sb.append("    firstName: ").append(toIndentedString(firstName)).append("\n");
+    sb.append("    lastName: ").append(toIndentedString(lastName)).append("\n");
+    sb.append("    phone: ").append(toIndentedString(phone)).append("\n");
+    sb.append("    role: ").append(toIndentedString(role)).append("\n");
+    sb.append("}");
+    return sb.toString();
+  }
+
+  /**
+   * Convert the given object to string with each line indented by 4 spaces
+   * (except the first line).
+   */
+  private String toIndentedString(java.lang.Object o) {
+    if (o == null) {
+      return "null";
+    }
+    return o.toString().replace("\n", "\n    ");
+  }
 }
